+++ tests/testthat/test-replace_emoticon.R
@@ -1,0 +1,7 @@
+context("Checking replace_emoticon")
+
+test_that("replace_emoticon ...",{
+
+
+})
+